iterate directly over dicts, not sets
This fixes the issue because it is specified for 3.6 and above. In
python 3.6+, sets are not ordered, but dictionaries are. This means
traversal will be stable across sessions, guaranteeing the same patches
are generated.

For python 3.5 and below, the stability of dictionary traversal is not a
guarantee, unfortunately. But if you use CPython, the reference
implementation of python, it does happen to be an implementation detail.
See https://docs.python.org/2/library/stdtypes.html#dict.items for more
details.