fix item_added logic
When an AddOperation would match a prior RemoveOperation, a
MoveOperation is inserted instead, and the RemoveOperation is deleted.
But when the RemoveOperation is deleted, subsequent operations that
depended on it are supposed to have their paths adjusted by
_on_undo_remove. Previously, _on_undo_remove was guarded by type(key) ==
int, which fails to consider cases where the key is a string but the
operation path still needs to be adjusted. The correct guard should be
type(op.pointer.to_last(self.dst_doc)[0]) == list, because
_on_undo_remove should apply whenever the item removed was contained in
a list.